% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_utils.R
\name{get_all_model_metadata}
\alias{get_all_model_metadata}
\title{Get metadata for all models}
\usage{
get_all_model_metadata()
}
\value{
Tibble with one row per model
}
\description{
Get metadata for all models
}
