% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_utils.R
\name{get_model_schema}
\alias{get_model_schema}
\title{Get schema for a specific model}
\usage{
get_model_schema(model_id)
}
\arguments{
\item{model_id}{Model identifier}
}
\value{
Tibble describing input schema
}
\description{
Get schema for a specific model
}
