% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_example_pesticide.R
\name{model_example_pesticide_metadata}
\alias{model_example_pesticide_metadata}
\title{Metadata for example pesticide risk model}
\usage{
model_example_pesticide_metadata()
}
\value{
Named list describing the model
}
\description{
Metadata for example pesticide risk model
}
