% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_example_pesticide.R
\name{model_example_pesticide_schema}
\alias{model_example_pesticide_schema}
\title{Input schema for example pesticide model}
\usage{
model_example_pesticide_schema()
}
\value{
Tibble describing required input fields
}
\description{
Input schema for example pesticide model
}
