% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{run_model}
\alias{run_model}
\title{Run a model by ID}
\usage{
run_model(model_id, data, params = list(), config = list())
}
\arguments{
\item{model_id}{Model identifier}

\item{data}{Input data frame}

\item{params}{Named list of model parameters}

\item{config}{Optional config list}
}
\value{
Model output tibble
}
\description{
Run a model by ID
}
