% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_utils.R
\name{write_input_template}
\alias{write_input_template}
\title{Write a CSV input template for a given model}
\usage{
write_input_template(model_id, path = NULL)
}
\arguments{
\item{model_id}{Model identifier}

\item{path}{File path to write to (defaults into data/input_templates/)}
}
\description{
Write a CSV input template for a given model
}
